#5 refactor files structure and create FastApi app
