--- app/models/node_model.py
@@ -1,8 +1,0 @@
-from pydantic import BaseModel
-
-from app.models.block_model import Block
-
-
-class Node(BaseModel):
-    blocks_array: list[Block]
-    block_index: int | None = None
